refactoring. using Blueprints.
