+++ webapp/admin/static/css/styles.css
@@ -1,0 +1,137 @@
+html, body {
+	width: 100%;
+	height: 100%;
+	margin: 0;
+	padding: 0;
+}
+
+.clear {clear: both;}
+
+.content {
+	margin: 10px;
+}
+
+.text_error {
+	color: #FF4040;
+}
+
+.wnd_dlg_back {
+	z-index: 9999;
+	position: absolute;
+	left: 0;
+	top: 0;
+	background: rgba(64, 65, 69, 0.5);
+	padding: 10px;
+	-webkit-border-radius: 20px;
+	-moz-border-radius: 20px;
+	border-radius: 20px;
+}
+.wnd_dlg_back.login_wnd {
+	position: absolute;
+	left: 50%;
+	top: 50%;
+	width: 345px;
+	height: 220px;
+	margin: -110px 0 0 -173px;
+	background: rgba(64, 65, 69, 0.2);
+}
+
+div.login {
+	width: 300px;
+	height: 175px;
+	padding: 20px;
+	color: #c39000;
+	background: #fafdd1;
+	border: 2px solid #d5d8b2;
+	-webkit-border-radius: 20px;
+	-moz-border-radius: 20px;
+	border-radius: 20px;
+}
+div.login p {
+	margin: 0 0 10px 0;
+}
+div.login p a {
+	color: #777;
+}
+div.login .title {
+	text-align: center;
+	color: #000;
+	font-size: 24px;
+	margin-bottom: 40px;
+}
+.login_input {
+	font-size: 16px;
+	width: 100%;
+	background: #041B40;
+	color: #fff;
+}
+.login_button {
+	color: #444;
+	margin-top: 10px;
+	font-size: 14px;
+	padding: 2px;
+	min-width: 200px
+}
+
+.form-label {
+	display: inline-block;
+	min-width: 100px;
+}
+
+.flash {
+	padding: 10px;
+	color: #444;
+	margin: -30px 0 10px 0;
+}
+.flash.success {
+	border: 1px solid #21DB56;
+	background: #AEFFC5;
+}
+.flash.error {
+	border: 1px solid #FF4343;
+	background: #FF9C9C;
+}
+
+ul.mainmenu {
+	list-style: none;
+	background: #fafdd1;
+	height: 40px;
+	margin: 0;
+	padding: 0;
+	color: #444;
+	font-size: 20px;
+	overflow: hidden;
+}
+ul.mainmenu li {
+	float: left;
+	margin: 8px 40px 0 50px;
+}
+ul.mainmenu li.last {float: right;}
+
+ul.mainmenu li a {
+	color: #444;
+	text-decoration: none;
+}
+ul.mainmenu li a:hover {
+	color: #777;
+}
+
+ul.list-posts {
+	list-style: none;
+	margin: 0;
+	padding: 0;
+	max-width: 600px;
+}
+ul.list-posts li {
+	margin: 20px 0 0 0;
+	border: 1px solid #eee;
+}
+ul.list-posts .title {
+	margin: 0;
+	padding: 5px;
+	background: #eee;
+}
+ul.list-posts .annonce {
+	margin: 0;
+	padding: 10px 5px 10px 5px;
+}
