+++ webapp/admin/templates/admin/base_admin.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <link type="text/css"
+          href="{{ url_for('.static', filename='css/styles.css') }}" rel="
+          stylesheet" />
+    <meta charset="UTF-8">
+    <title>{{ title }}</title>
+</head>
+<body>
+{% if menu -%}
+    <ul class="mainmenu">
+        {% for p in menu %}
+        <li><a href="{{ url_for(p.url) }}">{{ p.title}}</a></li>
+        {% endfor %}
+    </ul>
+{% endif %}
+<div class="content">
+    {% block content -%}
+    {% endblock %}
+</div>
+</body>
+</html>
